bug fixes in access/groups.py

+++ access/groups.py
@@ -563,11 +563,12 @@ def on_identity_url_changed(evt):
             latest_group_key_id = my_keys.latest_key_id(group_key_id)
             latest_group_path = os.path.join(groups_dir, latest_group_key_id)
             lg.info('going to rename rotated group key: %r -> %r' % (group_key_id, latest_group_key_id, ))
-            try:
-                os.rename(old_group_path, latest_group_path)
-            except:
-                lg.exc()
-                continue
+            if os.path.isfile(old_group_path):
+                try:
+                    os.rename(old_group_path, latest_group_path)
+                except:
+                    lg.exc()
+                    continue
             active_groups()[latest_group_key_id] = active_groups().pop(group_key_id)
             group_member.rotate_active_group_memeber(group_key_id, latest_group_key_id)
         gm = group_member.get_active_group_member(group_key_id)
@@ -582,12 +583,13 @@ def on_identity_url_changed(evt):
             latest_customer_dir = os.path.join(brokers_dir, latest_customer_id)
             lg.info('going to rename rotated customer id: %r -> %r' % (customer_id, latest_customer_id, ))
             old_customer_dir = os.path.join(brokers_dir, customer_id)
-            try:
-                bpio.move_dir_recursive(old_customer_dir, latest_customer_dir)
-                bpio.rmdir_recursive(old_customer_dir)
-            except:
-                lg.exc()
-                continue
+            if os.path.isdir(old_customer_dir):
+                try:
+                    bpio.move_dir_recursive(old_customer_dir, latest_customer_dir)
+                    bpio.rmdir_recursive(old_customer_dir)
+                except:
+                    lg.exc()
+                    continue
             known_brokers()[latest_customer_id] = known_brokers().pop(customer_id)
         for broker_pos, broker_id in enumerate(known_brokers(latest_customer_id)):
             if not broker_id:
@@ -598,11 +600,12 @@ def on_identity_url_changed(evt):
                 latest_broker_path = os.path.join(latest_customer_dir, latest_broker_id)
                 lg.info('going to rename rotated broker id: %r -> %r' % (broker_id, latest_broker_id, ))
                 old_broker_path = os.path.join(latest_customer_dir, broker_id)
-                try:
-                    os.rename(old_broker_path, latest_broker_path)
-                except:
-                    lg.exc()
-                    continue
+                if os.path.isfile(old_broker_path):
+                    try:
+                        os.rename(old_broker_path, latest_broker_path)
+                    except:
+                        lg.exc()
+                        continue
                 if latest_broker_id in known_brokers(latest_customer_id):
                     lg.warn('broker %r already exist' % latest_broker_id)
                     continue
